[#64] refactor : 무한 스크롤 hook으로 분리, 마이페이지에 적용
